+++ 42-Cursus/ft_printf/sources/ft_treat_hex.c
@@ -6,13 +6,13 @@
 /*   By: antmoren <antmoren@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/02 09:06:56 by antmoren          #+#    #+#             */
-/*   Updated: 2022/06/03 13:12:58 by antmoren         ###   ########.fr       */
+/*   Updated: 2022/06/06 12:56:36 by antmoren         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
 
-int	ft_puthexmin(unsigned long int number, int sub_cnt)
+int	ft_puthexmin(unsigned int number, int sub_cnt)
 {
 	char	*base;
 
@@ -30,7 +30,7 @@ int	ft_puthexmin(unsigned long int number, int sub_cnt)
 	return (sub_cnt);
 }
 
-int	ft_puthexmay(unsigned long int number, int sub_cnt)
+int	ft_puthexmay(unsigned int number, int sub_cnt)
 {
 	char	*base;
 
@@ -48,13 +48,13 @@ int	ft_puthexmay(unsigned long int number, int sub_cnt)
 	return (sub_cnt);
 }
 
-int	ft_treat_hex_min(unsigned long int number, int sub_cnt)
+int	ft_treat_hex_min(unsigned int number, int sub_cnt)
 {
 	sub_cnt = ft_puthexmin(number, sub_cnt);
 	return (sub_cnt);
 }
 
-int	ft_treat_hex_may(unsigned long int number, int sub_cnt)
+int	ft_treat_hex_may(unsigned int number, int sub_cnt)
 {
 	sub_cnt = ft_puthexmay(number, sub_cnt);
 	return (sub_cnt);
